Divide the remaining item-separable oversized modules (#701)
The last four files over the 1,000-line review budget that divide by item
ownership:

| file | lines | submodules | largest |
| --- | --- | --- | --- |
| wow-constants/opcodes.rs | 1,646 | server, client | 969 |
| wow-world/map_manager/runtime.rs | 1,342 | creature, manager, instance | 659 |
| wow-constants/item.rs | 1,298 | classification, flags, stats, results | 463 |
| wow-database/battle_pet/account_adapter.rs | 1,228 | persistence, rows, lock_broker | 681 |

opcodes.rs was two enums, 963 and 672 lines, sharing nothing but the file.
runtime.rs was three inherent impls of three different types. The physical
ratchet's cohesion reviews drop from 76 to 72.

One mechanic these needed that the earlier divisions did not: a `pub(super)`
member *inside* a moved impl block. The item scanner only sees top-level
items, so an impl moves whole and its members' visibility keywords move with
it - and `pub(super)` then means one module narrower than before. The six
such methods in map_manager/runtime.rs become
`pub(in crate::map_manager)`, which is the reach they had while the impl
lived in map_manager/runtime.rs; without that, map_manager/movement.rs can
no longer call them. E0624 named every one.

Item multiset identical per divided file (2, 44, 28 and 3 items). Crate tests
unchanged - wow-constants 5, wow-database 358, wow-world 3,822 - no test path
renamed, and no target gains a warning. Architecture check, self-test,
physical ratchet and hotspot ratchet pass with every hotspot at or below
baseline.

Closes #701
