+++ crates/wow-constants/src/opcodes/server.rs
@@ -1,684 +1,8 @@
-// Copyright (c) 2026 alseif0x
-// RustyCore — WoW WotLK 3.4.3 server in Rust
-// Based on TrinityCore protocol research (https://github.com/TrinityCore/TrinityCore)
-// Licensed under GPL v3 — https://www.gnu.org/licenses/gpl-3.0.html
+//! Server packets.
+//!
+//! Separated from opcodes.rs under #701.
 
-//! Client and Server opcodes for the WoW 3.4.3 protocol.
-
-use num_derive::{FromPrimitive, ToPrimitive};
-
-/// Opcodes sent from the client to the server.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, FromPrimitive, ToPrimitive)]
-#[repr(u32)]
-pub enum ClientOpcodes {
-    AcceptGuildInvite = 0x35fe,
-    AcceptSocialContract = 0x374d,
-    AcceptTrade = 0x315a,
-    AcceptWargameInvite = 0x35e1,
-    AccountNotificationAcknowledged = 0x373d,
-    ActivateTaxi = 0x34ab,
-    AddonList = 0x35d8,
-    AddBattlenetFriend = 0x365c,
-    AddFriend = 0x36d8,
-    AddIgnore = 0x36dc,
-    AddToy = 0x3299,
-    AdventureMapStartQuest = 0x32d9,
-    AlterAppearance = 0x34f5,
-    AreaSpiritHealerQuery = 0x34b0,
-    AreaSpiritHealerQueue = 0x34b1,
-    AreaTrigger = 0x31d6,
-    ArenaTeamAccept = 0x36b9,
-    ArenaTeamDecline = 0x36ba,
-    ArenaTeamDisband = 0x36bd,
-    ArenaTeamLeader = 0x36be,
-    ArenaTeamLeave = 0x36bb,
-    ArenaTeamRemove = 0x36bc,
-    ArenaTeamRoster = 0x36b8,
-    AssignEquipmentSetSpec = 0x3207,
-    AttackStop = 0x3256,
-    AttackSwing = 0x3255,
-    AuctionableTokenSell = 0x36ee,
-    AuctionableTokenSellAtMarketPrice = 0x36ef,
-    AuctionHelloRequest = 0x34ca,
-    AuctionListBidderItems = 0x34d0,
-    AuctionListItems = 0x34cd,
-    AuctionListOwnerItems = 0x34cf,
-    AuctionListPendingSales = 0x34d2,
-    AuctionPlaceBid = 0x34d1,
-    AuctionRemoveItem = 0x34cc,
-    AuctionReplicateItems = 0x34ce,
-    AuctionSellItem = 0x34cb,
-    AuthContinuedSession = 0x3766,
-    AuthSession = 0x3765,
-    AutobankItem = 0x3997,
-    AutostoreBankItem = 0x3996,
-    AutoEquipItem = 0x3998,
-    AutoEquipItemSlot = 0x399d,
-    AutoGuildBankItem = 0x34b6,
-    AutoStoreBagItem = 0x3999,
-    AutoStoreGuildBankItem = 0x34bf,
-    AzeriteEmpoweredItemSelectPower = 0x32f2,
-    AzeriteEmpoweredItemViewed = 0x32e0,
-    AzeriteEssenceActivateEssence = 0x32f4,
-    AzeriteEssenceUnlockMilestone = 0x32f3,
-    BankerActivate = 0x34b3,
-    BattlefieldLeave = 0x3175,
-    BattlefieldList = 0x3181,
-    BattlefieldPort = 0x3525,
-    BattlemasterHello = 0x32b1,
-    BattlemasterJoin = 0x3520,
-    BattlemasterJoinArena = 0x3521,
-    BattlemasterJoinSkirmish = 0x3522,
-    BattlenetChallengeResponse = 0x36db,
-    BattlenetRequest = 0x36fd,
-    BattlePayAckFailedResponse = 0x36d5,
-    BattlePayCancelOpenCheckout = 0x371b,
-    BattlePayConfirmPurchaseResponse = 0x36d4,
-    BattlePayDistributionAssignToTarget = 0x36cb,
-    BattlePayDistributionAssignVas = 0x3742,
-    BattlePayGetProductList = 0x36c4,
-    BattlePayGetPurchaseList = 0x36c5,
-    BattlePayOpenCheckout = 0x3714,
-    BattlePayRequestPriceInfo = 0x3710,
-    BattlePayStartPurchase = 0x36d3,
-    BattlePayStartVasPurchase = 0x36fa,
-    BattlePetClearFanfare = 0x3126,
-    BattlePetRequestJournal = 0x3625,
-    BattlePetRequestJournalLock = 0x3624,
-    BattlePetSetBattleSlot = 0x362e,
-    BattlePetSetFlags = 0x3631,
-    BattlePetSummon = 0x362a,
-    BattlePetUpdateDisplayNotify = 0x31e0,
-    BattlePetUpdateNotify = 0x31df,
-    BeginTrade = 0x3157,
-    BinderActivate = 0x34b2,
-    BlackMarketOpen = 0x352b,
-    BugReport = 0x3687,
-    BusyTrade = 0x3158,
-    BuyBackItem = 0x34a4,
-    BuyBankSlot = 0x34b4,
-    BuyItem = 0x34a3,
-    BuyStableSlot = 0x316b,
-    CalendarAddEvent = 0x367f,
-    CalendarCommunityInvite = 0x3673,
-    CalendarComplain = 0x367b,
-    CalendarCopyEvent = 0x367a,
-    CalendarEventSignUp = 0x367d,
-    CalendarGet = 0x3671,
-    CalendarGetEvent = 0x3672,
-    CalendarGetNumPending = 0x367c,
-    CalendarInvite = 0x3674,
-    CalendarModeratorStatus = 0x3678,
-    CalendarRemoveEvent = 0x3679,
-    CalendarRemoveInvite = 0x3675,
-    CalendarRsvp = 0x3676,
-    CalendarStatus = 0x3677,
-    CalendarUpdateEvent = 0x3680,
-    CancelAura = 0x31af,
-    CancelAutoRepeatSpell = 0x34e7,
-    CancelCast = 0x329f,
-    CancelChannelling = 0x326a,
-    CancelGrowthAura = 0x326f,
-    CancelMountAura = 0x327f,
-    CancelQueuedSpell = 0x3182,
-    CancelTempEnchantment = 0x34f2,
-    CancelTrade = 0x315c,
-    CanDuel = 0x3664,
-    CanRedeemTokenForBalance = 0x370f,
-    CastSpell = 0x329c,
-    ChangeBagSlotFlag = 0x32c8,
-    ChangeBankBagSlotFlag = 0x32c9,
-    ChangeRealmTicket = 0x3701,
-    ChangeSubGroup = 0x364f,
-    CharacterCheckUpgrade = 0x36ce,
-    CharacterRenameRequest = 0x36c9,
-    CharacterUpgradeManualUnrevokeRequest = 0x36cc,
-    CharacterUpgradeStart = 0x36cd,
-    CharCustomize = 0x368c,
-    CharDelete = 0x369d,
-    CharRaceOrFactionChange = 0x3692,
-    ChatAddonMessage = 0x37ee,
-    ChatAddonMessageWhisper = 0x37ef,
-    ChatChannelAnnouncements = 0x37e3,
-    ChatChannelBan = 0x37e1,
-    ChatChannelDeclineInvite = 0x37e6,
-    ChatChannelDisplayList = 0x37d6,
-    ChatChannelInvite = 0x37df,
-    ChatChannelKick = 0x37e0,
-    ChatChannelList = 0x37d5,
-    ChatChannelModerator = 0x37db,
-    ChatChannelOwner = 0x37d9,
-    ChatChannelPassword = 0x37d7,
-    ChatChannelSetOwner = 0x37d8,
-    ChatChannelSilenceAll = 0x37e4,
-    ChatChannelUnban = 0x37e2,
-    ChatChannelUnmoderator = 0x37dc,
-    ChatChannelUnsilenceAll = 0x37e5,
-    ChatJoinChannel = 0x37c8,
-    ChatLeaveChannel = 0x37c9,
-    ChatMessageAfk = 0x37d3,
-    ChatMessageChannel = 0x37cf,
-    ChatMessageDnd = 0x37d4,
-    ChatMessageEmote = 0x37e8,
-    ChatMessageGuild = 0x37d1,
-    ChatMessageInstanceChat = 0x37ec,
-    ChatMessageOfficer = 0x37d2,
-    ChatMessageParty = 0x37ea,
-    ChatMessageRaid = 0x37eb,
-    ChatMessageRaidWarning = 0x37ed,
-    ChatMessageSay = 0x37e7,
-    ChatMessageWhisper = 0x37d0,
-    ChatMessageYell = 0x37e9,
-    ChatRegisterAddonPrefixes = 0x37cd,
-    ChatReportFiltered = 0x37cc,
-    ChatReportIgnored = 0x37cb,
-    ChatUnregisterAllAddonPrefixes = 0x37ce,
-    CheckIsAdventureMapPoiValid = 0x3247,
-    ClearTradeItem = 0x315e,
-    CloseInteraction = 0x3493,
-    CloseQuestChoice = 0x32a2,
-    CloseTraitSystemInteraction = 0x330e,
-    ClubPresenceSubscribe = 0x36ff,
-    CollectionItemSetFavorite = 0x3634,
-    CommentatorEnable = 0x35f1,
-    CommentatorEnterInstance = 0x35f5,
-    CommentatorExitInstance = 0x35f6,
-    CommentatorGetMapInfo = 0x35f2,
-    CommentatorGetPlayerCooldowns = 0x35f4,
-    CommentatorGetPlayerInfo = 0x35f3,
-    CommentatorStartWargame = 0x35f0,
-    CommerceTokenGetCount = 0x36ec,
-    CommerceTokenGetLog = 0x36f6,
-    CommerceTokenGetMarketPrice = 0x36ed,
-    Complaint = 0x366e,
-    CompleteCinematic = 0x3545,
-    CompleteMovie = 0x34dd,
-    ConfirmBarbersChoice = 0x320e,
-    ConfirmRespecWipe = 0x320d,
-    ConnectToFailed = 0x35d4,
-    ConsumableTokenBuy = 0x36f1,
-    ConsumableTokenBuyAtMarketPrice = 0x36f2,
-    ConsumableTokenCanVeteranBuy = 0x36f0,
-    ConsumableTokenRedeem = 0x36f4,
-    ConsumableTokenRedeemConfirmation = 0x36f5,
-    ContributionLastUpdateRequest = 0x355b,
-    ConversationLineStarted = 0x3546,
-    ConvertRaid = 0x3651,
-    CreateCharacter = 0x3645,
-    DbQueryBulk = 0x35e5,
-    DeclineGuildInvites = 0x351d,
-    DeclinePetition = 0x3534,
-    DeleteEquipmentSet = 0x350a,
-    DelFriend = 0x36d9,
-    DelIgnore = 0x36dd,
-    DestroyItem = 0x3293,
-    DfBootPlayerVote = 0x3618,
-    DfGetJoinStatus = 0x3616,
-    DfGetSystemInfo = 0x3615,
-    DfJoin = 0x360b,
-    DfLeave = 0x3614,
-    DfProposalResponse = 0x3609,
-    DfReadyCheckResponse = 0x361c,
-    DfSetRoles = 0x3617,
-    DfTeleport = 0x3619,
-    DiscardedTimeSyncAcks = 0x3a41,
-    DismissCritter = 0x34f9,
-    DoReadyCheck = 0x3635,
-    DuelResponse = 0x34e2,
-    EjectPassenger = 0x323c,
-    Emote = 0x3541,
-    EnableNagle = 0x376b,
-    EnableTaxiNode = 0x34a9,
-    EngineSurvey = 0x36eb,
-    EnterEncryptedModeAck = 0x3767,
-    EnumCharacters = 0x35e9,
-    EnumCharactersDeletedByClient = 0x36e5,
-    FarSight = 0x34e8,
-    GameEventDebugDisable = 0x31b2,
-    GameEventDebugEnable = 0x31b1,
-    GameObjReportUse = 0x34ef,
-    GameObjUse = 0x34ee,
-    GenerateRandomCharacterName = 0x35e8,
-    GetAccountCharacterList = 0x36bf,
-    GetAccountNotifications = 0x373c,
-    GetItemPurchaseData = 0x352f,
-    GetMirrorImageData = 0x3297,
-    GetRemainingGameTime = 0x36f3,
-    GetUndeleteCharacterCooldownStatus = 0x36e7,
-    GetVasAccountCharacterList = 0x36f8,
-    GetVasTransferTargetRealmList = 0x36f9,
-    GmTicketAcknowledgeSurvey = 0x3690,
-    GmTicketGetCaseStatus = 0x368f,
-    GmTicketGetSystemStatus = 0x368e,
-    GossipSelectOption = 0x3494,
-    GuildAddBattlenetFriend = 0x308c,
-    GuildAddRank = 0x3064,
-    GuildAssignMemberRank = 0x305f,
-    GuildAutoDeclineInvitation = 0x3061,
-    GuildBankActivate = 0x34b5,
-    GuildBankBuyTab = 0x34c3,
-    GuildBankDepositMoney = 0x34c5,
-    GuildBankLogQuery = 0x3082,
-    GuildBankQueryTab = 0x34c2,
-    GuildBankRemainingWithdrawMoneyQuery = 0x3083,
-    GuildBankSetTabText = 0x3086,
-    GuildBankTextQuery = 0x3087,
-    GuildBankUpdateTab = 0x34c4,
-    GuildBankWithdrawMoney = 0x34c6,
-    GuildChangeNameRequest = 0x307e,
-    GuildDeclineInvitation = 0x3060,
-    GuildDelete = 0x3068,
-    GuildDeleteRank = 0x3065,
-    GuildDemoteMember = 0x305e,
-    GuildEventLogQuery = 0x3085,
-    GuildGetRanks = 0x306d,
-    GuildGetRoster = 0x3073,
-    GuildInviteByName = 0x3608,
-    GuildLeave = 0x3062,
-    GuildOfficerRemoveMember = 0x3063,
-    GuildPermissionsQuery = 0x3084,
-    GuildPromoteMember = 0x305d,
-    GuildQueryMembersForRecipe = 0x306b,
-    GuildReplaceGuildMaster = 0x3088,
-    GuildSetAchievementTracking = 0x306f,
-    GuildSetGuildMaster = 0x36d0,
-    GuildSetMemberNote = 0x3072,
-    GuildSetRankPermissions = 0x3067,
-    GuildShiftRank = 0x3066,
-    GuildUpdateInfoText = 0x3075,
-    GuildUpdateMotdText = 0x3074,
-    HearthAndResurrect = 0x3506,
-    HotfixRequest = 0x35e6,
-    IgnoreTrade = 0x3159,
-    InitiateRolePoll = 0x35da,
-    InitiateTrade = 0x3156,
-    Inspect = 0x3529,
-    InspectPvp = 0x36a3,
-    InstanceLockResponse = 0x350b,
-    ItemPurchaseRefund = 0x3530,
-    ItemTextQuery = 0x32c5,
-    KeepAlive = 0x3681,
-    KeyboundOverride = 0x3223,
-    LatencyReport = 0x3771,
-    LearnPreviewTalents = 0x3553,
-    LearnPreviewTalentsPet = 0x3555,
-    LearnTalent = 0x3552,
-    LeaveGroup = 0x364c,
-    LfgListApplyToGroup = 0x360f,
-    LfgListCancelApplication = 0x3610,
-    LfgListDeclineApplicant = 0x3611,
-    LfgListGetStatus = 0x360d,
-    LfgListInviteApplicant = 0x3612,
-    LfgListInviteResponse = 0x3613,
-    LfgListJoin = 0x32f0,
-    LfgListLeave = 0x360c,
-    LfgListSearch = 0x360e,
-    LfgListSetRoles = 0x3302,
-    LfgListUpdateRequest = 0x32f1,
-    ListInventory = 0x34a1,
-    LiveRegionAccountRestore = 0x36c2,
-    LiveRegionCharacterCopy = 0x36c1,
-    LiveRegionGetAccountCharacterList = 0x36c0,
-    LoadingScreenNotify = 0x35f9,
-    LogoutCancel = 0x34d8,
-    LogoutInstant = 0x34d9,
-    LogoutRequest = 0x34d6,
-    LogDisconnect = 0x3769,
-    LogStreamingError = 0x376d,
-    LootItem = 0x3211,
-    LootMoney = 0x3210,
-    LootRelease = 0x3213,
-    LootRoll = 0x3214,
-    LootUnit = 0x320f,
-    SetLootSpecialization = 0xbadd,
-    LowLevelRaid1 = 0x36a1,
-    LowLevelRaid2 = 0x3512,
-    MailCreateTextItem = 0x353b,
-    MailDelete = 0x3225,
-    MailGetList = 0x3536,
-    MailMarkAsRead = 0x353a,
-    MailReturnToSender = 0x3658,
-    MailTakeItem = 0x3538,
-    MailTakeMoney = 0x3537,
-    MasterLootItem = 0x3212,
-    MergeGuildBankItemWithGuildBankItem = 0x34c0,
-    MergeGuildBankItemWithItem = 0x34bd,
-    MergeItemWithGuildBankItem = 0x34bb,
-    MinimapPing = 0x364e,
-    MissileTrajectoryCollision = 0x318d,
-    MountClearFanfare = 0x3127,
-    MountSetFavorite = 0x3633,
-    MountSpecialAnim = 0x3280,
-    MoveAddImpulseAck = 0x3a50,
-    MoveApplyInertiaAck = 0x3a4e,
-    MoveApplyMovementForceAck = 0x3a15,
-    MoveChangeTransport = 0x3a2f,
-    MoveChangeVehicleSeats = 0x3a34,
-    MoveCollisionDisableAck = 0x3a39,
-    MoveCollisionEnableAck = 0x3a3a,
-    MoveDismissVehicle = 0x3a33,
-    MoveDoubleJump = 0x39eb,
-    MoveEnableDoubleJumpAck = 0x3a1e,
-    MoveEnableSwimToFlyTransAck = 0x3a24,
-    MoveFallLand = 0x39fb,
-    MoveFallReset = 0x3a19,
-    MoveFeatherFallAck = 0x3a1c,
-    MoveForceFlightBackSpeedChangeAck = 0x3a2e,
-    MoveForceFlightSpeedChangeAck = 0x3a2d,
-    MoveForcePitchRateChangeAck = 0x3a32,
-    MoveForceRootAck = 0x3a0e,
-    MoveForceRunBackSpeedChangeAck = 0x3a0c,
-    MoveForceRunSpeedChangeAck = 0x3a0b,
-    MoveForceSwimBackSpeedChangeAck = 0x3a22,
-    MoveForceSwimSpeedChangeAck = 0x3a0d,
-    MoveForceTurnRateChangeAck = 0x3a23,
-    MoveForceUnrootAck = 0x3a0f,
-    MoveForceWalkSpeedChangeAck = 0x3a21,
-    MoveGravityDisableAck = 0x3a35,
-    MoveGravityEnableAck = 0x3a36,
-    MoveGuildBankItem = 0x34ba,
-    MoveHeartbeat = 0x3a10,
-    MoveHoverAck = 0x3a13,
-    MoveInertiaDisableAck = 0x3a37,
-    MoveInertiaEnableAck = 0x3a38,
-    MoveInitActiveMoverComplete = 0x3a46,
-    MoveJump = 0x39ea,
-    MoveKnockBackAck = 0x3a12,
-    MoveRemoveInertiaAck = 0x3a4f,
-    MoveRemoveMovementForces = 0x3a17,
-    MoveRemoveMovementForceAck = 0x3a16,
-    MoveSeamlessTransferComplete = 0x3a44,
-    MoveSetAdvFly = 0x3a52,
-    MoveSetAdvFlyingAddImpulseMaxSpeedAck = 0x3a58,
-    MoveSetAdvFlyingAirFrictionAck = 0x3a53,
-    MoveSetAdvFlyingBankingRateAck = 0x3a59,
-    MoveSetAdvFlyingDoubleJumpVelModAck = 0x3a56,
-    MoveSetAdvFlyingGlideStartMinHeightAck = 0x3a57,
-    MoveSetAdvFlyingLaunchSpeedCoefficientAck = 0x3a60,
-    MoveSetAdvFlyingLiftCoefficientAck = 0x3a55,
-    MoveSetAdvFlyingMaxVelAck = 0x3a54,
-    MoveSetAdvFlyingOverMaxDecelerationAck = 0x3a5e,
-    MoveSetAdvFlyingPitchingRateDownAck = 0x3a5a,
-    MoveSetAdvFlyingPitchingRateUpAck = 0x3a5b,
-    MoveSetAdvFlyingSurfaceFrictionAck = 0x3a5d,
-    MoveSetAdvFlyingTurnVelocityThresholdAck = 0x3a5c,
-    MoveSetCanAdvFlyAck = 0x3a51,
-    MoveSetCanFlyAck = 0x3a27,
-    MoveSetCanTurnWhileFallingAck = 0x3a25,
-    MoveSetCollisionHeightAck = 0x3a3b,
-    MoveSetFacing = 0x3a09,
-    MoveSetFacingHeartbeat = 0x3a5f,
-    MoveSetFly = 0x3a28,
-    MoveSetIgnoreMovementForcesAck = 0x3a26,
-    MoveSetModMovementForceMagnitudeAck = 0x3a42,
-    MoveSetPitch = 0x3a0a,
-    MoveSetRunMode = 0x39f2,
-    MoveSetVehicleRecIdAck = 0x3a14,
-    MoveSetWalkMode = 0x39f3,
-    MoveSplineDone = 0x3a18,
-    MoveStartAscend = 0x3a29,
-    MoveStartBackward = 0x39e5,
-    MoveStartDescend = 0x3a30,
-    MoveStartForward = 0x39e4,
-    MoveStartPitchDown = 0x39f0,
-    MoveStartPitchUp = 0x39ef,
-    MoveStartStrafeLeft = 0x39e7,
-    MoveStartStrafeRight = 0x39e8,
-    MoveStartSwim = 0x39fc,
-    MoveStartTurnLeft = 0x39ec,
-    MoveStartTurnRight = 0x39ed,
-    MoveStop = 0x39e6,
-    MoveStopAscend = 0x3a2a,
-    MoveStopPitch = 0x39f1,
-    MoveStopStrafe = 0x39e9,
-    MoveStopSwim = 0x39fd,
-    MoveStopTurn = 0x39ee,
-    MoveTeleportAck = 0x39fa,
-    MoveTimeSkipped = 0x3a1b,
-    MoveUpdateFallSpeed = 0x3a1a,
-    MoveWaterWalkAck = 0x3a1d,
-    NextCinematicCamera = 0x3544,
-    ObjectUpdateFailed = 0x3183,
-    ObjectUpdateRescued = 0x3184,
-    OfferPetition = 0x32fd,
-    OpeningCinematic = 0x3543,
-    OpenItem = 0x32c6,
-    OptOutOfLoot = 0x34f6,
-    OverrideScreenFlash = 0x351e,
-    PartyInvite = 0x3604,
-    PartyInviteResponse = 0x3606,
-    PartyUninvite = 0x364a,
-    PetitionBuy = 0x34c8,
-    PetitionRenameGuild = 0x36d1,
-    PetitionShowList = 0x34c7,
-    PetitionShowSignatures = 0x34c9,
-    PetAbandon = 0x348d,
-    PetAction = 0x348b,
-    PetCancelAura = 0x348e,
-    PetCastSpell = 0x329b,
-    PetLearnTalent = 0x3554,
-    PetRename = 0x3686,
-    PetSetAction = 0x348a,
-    PetSpellAutocast = 0x348f,
-    PetStopAttack = 0x348c,
-    Ping = 0x3768,
-    PlayerLogin = 0x35eb,
-    PushQuestToParty = 0x349f,
-    PvpLogData = 0x317f,
-    QueryArenaTeam = 0x36a4,
-    QueryBattlePetName = 0x3276,
-    QueryCorpseLocationFromClient = 0x3662,
-    QueryCorpseTransport = 0x3663,
-    QueryCountdownTimer = 0x31aa,
-    QueryCreature = 0x3270,
-    QueryGameObject = 0x3271,
-    QueryGuildInfo = 0x368b,
-    QueryInspectAchievements = 0x3500,
-    QueryNextMailTime = 0x3539,
-    QueryNpcText = 0x3272,
-    QueryPageText = 0x3274,
-    QueryPetition = 0x3277,
-    QueryPetName = 0x3275,
-    QueryPlayerNames = 0x3772,
-    QueryPlayerNamesForCommunity = 0x3770,
-    QueryPlayerNameByCommunityId = 0x376f,
-    QueryQuestCompletionNpcs = 0x3177,
-    QueryQuestInfo = 0x3273,
-    QueryQuestItemUsability = 0x3178,
-    QueryRealmName = 0x368a,
-    QueryTime = 0x34d5,
-    QueryTreasurePicker = 0x32dc,
-    QueryVoidStorage = 0x31a3,
-    QuestConfirmAccept = 0x349e,
-    QuestGiverAcceptQuest = 0x3498,
-    QuestGiverChooseReward = 0x349a,
-    QuestGiverCloseQuest = 0x3549,
-    QuestGiverCompleteQuest = 0x3499,
-    QuestGiverHello = 0x3496,
-    QuestGiverQueryQuest = 0x3497,
-    QuestGiverRequestReward = 0x349b,
-    QuestGiverStatusMultipleQuery = 0x349d,
-    QuestGiverStatusQuery = 0x349c,
-    QuestGiverStatusTrackedQuery = 0x356b,
-    QuestLogRemoveQuest = 0x352e,
-    QuestPoiQuery = 0x36b2,
-    QuestPushResult = 0x34a0,
-    QueuedMessagesEnd = 0x376c,
-    QuickJoinAutoAcceptRequests = 0x370d,
-    QuickJoinRequestInvite = 0x370c,
-    QuickJoinRespondToInvite = 0x370b,
-    RandomRoll = 0x3657,
-    ReadyCheckResponse = 0x3636,
-    ReadItem = 0x32c7,
-    ReclaimCorpse = 0x34db,
-    RemoveGlyph = 0x3300,
-    RemoveNewItem = 0x32df,
-    ReorderCharacters = 0x35ea,
-    RepairItem = 0x34ec,
-    RepopRequest = 0x3526,
-    ReportClientVariables = 0x3707,
-    ReportEnabledAddons = 0x3706,
-    ReportFrozenWhileLoadingMap = 0x36aa,
-    ReportKeybindingExecutionCounts = 0x3708,
-    ReportPvpPlayerAfk = 0x34f4,
-    ReportServerLag = 0x32fb,
-    RequestAccountData = 0x3694,
-    RequestAreaPoiUpdate = 0x32de,
-    RequestBattlefieldStatus = 0x35dd,
-    RequestCemeteryList = 0x3179,
-    RequestConquestFormulaConstants = 0x32b4,
-    RequestCrowdControlSpell = 0x352a,
-    RequestForcedReactions = 0x3205,
-    RequestHonorStats = 0x317e,
-    RequestLfgListBlacklist = 0x32a4,
-    RequestPartyJoinUpdates = 0x35f8,
-    RequestPartyMemberStats = 0x3656,
-    RequestPetInfo = 0x3490,
-    RequestPlayedTime = 0x327a,
-    RequestPvpRewards = 0x3196,
-    RequestRaidInfo = 0x36d2,
-    RequestRatedPvpInfo = 0x35e4,
-    RequestScheduledPvpInfo = 0x3197,
-    RequestStabledPets = 0x3491,
-    RequestVehicleExit = 0x3237,
-    RequestVehicleNextSeat = 0x3239,
-    RequestVehiclePrevSeat = 0x3238,
-    RequestVehicleSwitchSeat = 0x323a,
-    RequestWorldQuestUpdate = 0x32dd,
-    ResetInstances = 0x366a,
-    ResurrectResponse = 0x3685,
-    RideVehicleInteract = 0x323b,
-    SaveAccountDataExport = 0x3751,
-    SaveCufProfiles = 0x318e,
-    SaveEquipmentSet = 0x3509,
-    SaveGuildEmblem = 0x32a8,
-    ScenePlaybackCanceled = 0x3220,
-    ScenePlaybackComplete = 0x321f,
-    SceneTriggerEvent = 0x3221,
-    SelfRes = 0x3531,
-    SellItem = 0x34a2,
-    SendContactList = 0x36d7,
-    SendMail = 0x35fb,
-    SendTextEmote = 0x3488,
-    ServerTimeOffsetRequest = 0x369c,
-    SetActionBarToggles = 0x3532,
-    SetActionButton = 0x355d,
-    SetActiveMover = 0x3a3c,
-    SetAdvancedCombatLogging = 0x32b5,
-    SetAmmo = 0x355e,
-    SetAssistantLeader = 0x3652,
-    SetContactNotes = 0x36da,
-    SetCurrencyFlags = 0x316c,
-    SetDifficultyId = 0x3222,
-    SetDungeonDifficulty = 0x3684,
-    SetEveryoneIsAssistant = 0x361a,
-    SetFactionAtWar = 0x34de,
-    SetFactionInactive = 0x34e0,
-    SetFactionNotAtWar = 0x34df,
-    SetGameEventDebugViewState = 0x31b9,
-    SetInsertItemsLeftToRight = 0x32ca,
-    SetLootMethod = 0x364b,
-    SetPartyAssignment = 0x3654,
-    SetPartyLeader = 0x364d,
-    SetPlayerDeclinedNames = 0x3689,
-    SetPreferredCemetery = 0x317a,
-    SetPvp = 0x32ac,
-    SetRaidDifficulty = 0x36e3,
-    SetRole = 0x35d9,
-    SetSelection = 0x3528,
-    SetSheathed = 0x3489,
-    SetTaxiBenchmarkMode = 0x34f3,
-    SetTitle = 0x327e,
-    SetTradeGold = 0x315f,
-    SetTradeItem = 0x315d,
-    SetTradeSpell = 0x3160,
-    SetWatchedFaction = 0x34e1,
-    ShowingCloak = 0x3569,
-    ShowingHelm = 0x3568,
-    ShowTradeSkill = 0x36ca,
-    SignPetition = 0x3533,
-    SilencePartyTalker = 0x3655,
-    SocialContractRequest = 0x374c,
-    SocketGems = 0x34eb,
-    SpawnTrackingUpdate = 0x3292,
-    SpellClick = 0x3495,
-    SpiritHealerActivate = 0x34af,
-    SplitGuildBankItem = 0x34c1,
-    SplitGuildBankItemToInventory = 0x34be,
-    SplitItem = 0x399c,
-    SplitItemToGuildBank = 0x34bc,
-    StablePet = 0x3168,
-    StableSwapPet = 0x316a,
-    StandStateChange = 0x318c,
-    StartSpectatorWarGame = 0x35e0,
-    StartWarGame = 0x35df,
-    StoreGuildBankItem = 0x34b7,
-    SubmitUserFeedback = 0x3693,
-    SummonResponse = 0x366c,
-    SupportTicketSubmitBug = 0x3648,
-    SupportTicketSubmitComplaint = 0x3647,
-    SupportTicketSubmitSuggestion = 0x3649,
-    SuspendCommsAck = 0x3764,
-    SuspendTokenResponse = 0x376a,
-    SwapGuildBankItemWithGuildBankItem = 0x34b9,
-    SwapInvItem = 0x399b,
-    SwapItem = 0x399a,
-    SwapItemWithGuildBankItem = 0x34b8,
-    SwapSubGroups = 0x3650,
-    SwapVoidItem = 0x31a5,
-    TabardVendorActivate = 0x32a9,
-    TalkToGossip = 0x3492,
-    TaxiNodeStatusQuery = 0x34a8,
-    TaxiQueryAvailableNodes = 0x34aa,
-    TaxiRequestEarlyLanding = 0x34ac,
-    TimeAdjustmentResponse = 0x3a40,
-    TimeSyncResponse = 0x3a3d,
-    TimeSyncResponseDropped = 0x3a3f,
-    TimeSyncResponseFailed = 0x3a3e,
-    ToggleDifficulty = 0x3659,
-    TogglePvp = 0x32ab,
-    TotemDestroyed = 0x34f8,
-    ToyClearFanfare = 0x3128,
-    TrainerBuySpell = 0x34ae,
-    TrainerList = 0x34ad,
-    TraitsCommitConfig = 0x3303,
-    TurnInPetition = 0x3535,
-    Tutorial = 0x36e4,
-    UnacceptTrade = 0x315b,
-    UndeleteCharacter = 0x36e6,
-    UnlearnSkill = 0x34e5,
-    UnlockVoidStorage = 0x31a2,
-    UnstablePet = 0x3169,
-    UpdateAadcStatus = 0x3741,
-    UpdateAccountData = 0x3695,
-    UpdateAreaTriggerVisual = 0x329e,
-    UpdateClientSettings = 0x3666,
-    UpdateMissileTrajectory = 0x3a43,
-    UpdateRaidTarget = 0x3653,
-    UpdateSpellVisual = 0x329d,
-    UpdateVasPurchaseStates = 0x36fb,
-    UsedFollow = 0x3189,
-    UseCritterItem = 0x3241,
-    UseEquipmentSet = 0x3995,
-    UseItem = 0x3298,
-    UseToy = 0x329a,
-    VasCheckTransferOk = 0x3713,
-    VasGetQueueMinutes = 0x3712,
-    VasGetServiceStatus = 0x3711,
-    ViolenceLevel = 0x3187,
-    VoiceChannelSttTokenRequest = 0x3717,
-    VoiceChatJoinChannel = 0x3718,
-    VoiceChatLogin = 0x3716,
-    VoidStorageTransfer = 0x31a4,
-    Warden3Data = 0x35ed,
-    Who = 0x3683,
-    WhoIs = 0x3682,
-    WorldPortResponse = 0x35fa,
-    WrapItem = 0x3994,
-
-    Max = 0x3FFF,
-    None = 0,
-}
+use super::*;
 
 /// Opcodes sent from the server to the client.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, FromPrimitive, ToPrimitive)]
