+++ crates/wow-database/src/battle_pet/account_adapter/persistence.rs
@@ -1,284 +1,13 @@
-//! MariaDB adapter for account-scoped battle-pet durability.
+//! Persistence packets.
+//!
+//! Separated from account_adapter.rs under #701.
 
-use std::collections::HashSet;
-use std::sync::{
-    Arc,
-    atomic::{AtomicU64, Ordering},
-};
-use std::time::Duration;
-
-use sqlx::{MySql, Row, Transaction};
-use tokio::{
-    sync::{mpsc, oneshot},
-    time::MissedTickBehavior,
-};
-use wow_persistence::{
-    BATTLE_PET_GUID_COUNTER_LIMIT_LIKE_CPP,
-    BattlePetAccountPersistencePortLikeCpp as BattlePetPersistenceLikeCpp,
-    BattlePetAddRequestKeyLikeCpp, BattlePetDeclinedNamesLikeCpp, BattlePetPersistenceErrorLikeCpp,
-    BattlePetProcessLeaseLikeCpp, DurableBattlePetAddLikeCpp, DurableBattlePetAddReceiptLikeCpp,
-    DurableBattlePetRowLikeCpp, DurableBattlePetSlotLikeCpp, LoadedBattlePetAccountLikeCpp,
-    PersistBattlePetAddOutcomeLikeCpp, PersistenceFutureLikeCpp as PersistenceFuture,
-};
-
-use crate::{
-    DatabaseError, LoginDatabase, LoginStatements, SqlResult, SqlTransaction,
-    SqlTransactionCommitError,
-};
-
-const BATTLE_PET_PROCESS_LEASE_VERIFY_INTERVAL_LIKE_CPP: Duration = Duration::from_secs(30);
+use super::*;
 
 #[derive(Debug)]
 pub struct LoginBattlePetPersistenceLikeCpp {
     db: Arc<LoginDatabase>,
     lock_broker: BattlePetAccountLockBrokerLikeCpp,
-}
-
-#[derive(Debug, Clone)]
-struct BattlePetAccountLockBrokerLikeCpp {
-    commands: mpsc::UnboundedSender<BattlePetAccountLockCommandLikeCpp>,
-    epoch: Arc<AtomicU64>,
-}
-
-enum BattlePetAccountLockCommandLikeCpp {
-    Acquire {
-        account_id: u32,
-        result: oneshot::Sender<Result<Option<(String, u64)>, String>>,
-    },
-    Release {
-        lock_name: String,
-        epoch: u64,
-    },
-}
-
-struct LoginBattlePetProcessLeaseLikeCpp {
-    lock_broker: BattlePetAccountLockBrokerLikeCpp,
-    lock_name: Option<String>,
-    broker_epoch: u64,
-    fence: u64,
-}
-
-struct BattlePetBrokerLeaseReservationLikeCpp {
-    lock_broker: BattlePetAccountLockBrokerLikeCpp,
-    lock_name: Option<String>,
-    broker_epoch: u64,
-}
-
-impl Drop for BattlePetBrokerLeaseReservationLikeCpp {
-    fn drop(&mut self) {
-        if let Some(lock_name) = self.lock_name.take() {
-            self.lock_broker
-                .release_like_cpp(lock_name, self.broker_epoch);
-        }
-    }
-}
-
-impl BattlePetProcessLeaseLikeCpp for LoginBattlePetProcessLeaseLikeCpp {
-    fn is_valid_like_cpp(&self) -> bool {
-        !self.lock_broker.commands.is_closed()
-            && self.lock_broker.epoch.load(Ordering::Acquire) == self.broker_epoch
-    }
-
-    fn fence_like_cpp(&self) -> u64 {
-        self.fence
-    }
-}
-
-impl Drop for LoginBattlePetProcessLeaseLikeCpp {
-    fn drop(&mut self) {
-        if let Some(lock_name) = self.lock_name.take() {
-            let _ = self
-                .lock_broker
-                .commands
-                .send(BattlePetAccountLockCommandLikeCpp::Release {
-                    lock_name,
-                    epoch: self.broker_epoch,
-                });
-        }
-    }
-}
-
-impl BattlePetAccountLockBrokerLikeCpp {
-    fn spawn_like_cpp(db: Arc<LoginDatabase>) -> Self {
-        let (commands, receiver) = mpsc::unbounded_channel();
-        let epoch = Arc::new(AtomicU64::new(1));
-        tokio::spawn(run_battle_pet_account_lock_broker_like_cpp(
-            db,
-            receiver,
-            Arc::clone(&epoch),
-        ));
-        Self { commands, epoch }
-    }
-
-    async fn acquire_like_cpp(
-        &self,
-        account_id: u32,
-    ) -> Result<Option<(String, u64)>, BattlePetPersistenceErrorLikeCpp> {
-        let (result, response) = oneshot::channel();
-        self.commands
-            .send(BattlePetAccountLockCommandLikeCpp::Acquire { account_id, result })
-            .map_err(|_| {
-                BattlePetPersistenceErrorLikeCpp::Database(
-                    "battle-pet account lock broker stopped".to_string(),
-                )
-            })?;
-        response
-            .await
-            .map_err(|_| {
-                BattlePetPersistenceErrorLikeCpp::Database(
-                    "battle-pet account lock broker dropped acquisition".to_string(),
-                )
-            })?
-            .map_err(BattlePetPersistenceErrorLikeCpp::Database)
-    }
-
-    fn release_like_cpp(&self, lock_name: String, epoch: u64) {
-        let _ = self
-            .commands
-            .send(BattlePetAccountLockCommandLikeCpp::Release { lock_name, epoch });
-    }
-}
-
-async fn open_battle_pet_lock_broker_connection_like_cpp(
-    db: &LoginDatabase,
-) -> Result<(sqlx::MySqlConnection, String), String> {
-    let pooled = db
-        .pool()
-        .acquire()
-        .await
-        .map_err(|error| error.to_string())?;
-    let mut connection = pooled.detach();
-    let database_scope: String =
-        sqlx::query_scalar("SELECT LEFT(SHA2(COALESCE(DATABASE(), ''), 256), 32)")
-            .fetch_one(&mut connection)
-            .await
-            .map_err(|error| error.to_string())?;
-    Ok((connection, database_scope))
-}
-
-fn invalidate_battle_pet_lock_broker_like_cpp(
-    connection: &mut Option<(sqlx::MySqlConnection, String)>,
-    held_locks: &mut HashSet<String>,
-    epoch: &AtomicU64,
-) {
-    *connection = None;
-    held_locks.clear();
-    epoch.fetch_add(1, Ordering::AcqRel);
-}
-
-fn battle_pet_account_lock_name_like_cpp(database_scope: &str, account_id: u32) -> String {
-    format!("rustycore:bp:{database_scope}:{account_id}")
-}
-
-async fn run_battle_pet_account_lock_broker_like_cpp(
-    db: Arc<LoginDatabase>,
-    mut commands: mpsc::UnboundedReceiver<BattlePetAccountLockCommandLikeCpp>,
-    epoch: Arc<AtomicU64>,
-) {
-    let mut connection: Option<(sqlx::MySqlConnection, String)> = None;
-    let mut held_locks = HashSet::new();
-    let mut verify_interval =
-        tokio::time::interval(BATTLE_PET_PROCESS_LEASE_VERIFY_INTERVAL_LIKE_CPP);
-    verify_interval.set_missed_tick_behavior(MissedTickBehavior::Skip);
-    loop {
-        tokio::select! {
-            biased;
-            command = commands.recv() => {
-                let Some(command) = command else { return; };
-                match command {
-                    BattlePetAccountLockCommandLikeCpp::Acquire { account_id, result } => {
-                        if connection.is_none() {
-                            match open_battle_pet_lock_broker_connection_like_cpp(&db).await {
-                                Ok(opened) => connection = Some(opened),
-                                Err(error) => {
-                                    let _ = result.send(Err(error));
-                                    continue;
-                                }
-                            }
-                        }
-                        let (_, database_scope) = connection.as_ref().expect("broker connection opened");
-                        let lock_name = battle_pet_account_lock_name_like_cpp(database_scope, account_id);
-                        if held_locks.contains(&lock_name) {
-                            let _ = result.send(Ok(None));
-                            continue;
-                        }
-                        let acquired = {
-                            let (connection, _) = connection.as_mut().expect("broker connection opened");
-                            sqlx::query_scalar::<_, Option<i64>>("SELECT GET_LOCK(?, 0)")
-                                .bind(&lock_name)
-                                .fetch_one(connection)
-                                .await
-                        };
-                        match acquired {
-                            Ok(Some(1)) => {
-                                held_locks.insert(lock_name.clone());
-                                let lease_epoch = epoch.load(Ordering::Acquire);
-                                if result.send(Ok(Some((lock_name.clone(), lease_epoch)))).is_err() {
-                                    held_locks.remove(&lock_name);
-                                    let release = {
-                                        let (connection, _) = connection.as_mut().expect("broker connection opened");
-                                        sqlx::query_scalar::<_, Option<i64>>("SELECT RELEASE_LOCK(?)")
-                                            .bind(&lock_name)
-                                            .fetch_one(connection)
-                                            .await
-                                    };
-                                    if release.is_err() {
-                                        invalidate_battle_pet_lock_broker_like_cpp(
-                                            &mut connection,
-                                            &mut held_locks,
-                                            &epoch,
-                                        );
-                                    }
-                                }
-                            }
-                            Ok(_) => { let _ = result.send(Ok(None)); }
-                            Err(error) => {
-                                invalidate_battle_pet_lock_broker_like_cpp(
-                                    &mut connection,
-                                    &mut held_locks,
-                                    &epoch,
-                                );
-                                let _ = result.send(Err(error.to_string()));
-                            }
-                        }
-                    }
-                    BattlePetAccountLockCommandLikeCpp::Release { lock_name, epoch: lease_epoch } => {
-                        if lease_epoch != epoch.load(Ordering::Acquire) || !held_locks.remove(&lock_name) {
-                            continue;
-                        }
-                        let release = if let Some((connection, _)) = connection.as_mut() {
-                            sqlx::query_scalar::<_, Option<i64>>("SELECT RELEASE_LOCK(?)")
-                                .bind(&lock_name)
-                                .fetch_one(connection)
-                                .await
-                        } else {
-                            continue;
-                        };
-                        if release.is_err() {
-                            invalidate_battle_pet_lock_broker_like_cpp(
-                                &mut connection,
-                                &mut held_locks,
-                                &epoch,
-                            );
-                        }
-                    }
-                }
-            }
-            _ = verify_interval.tick(), if connection.is_some() => {
-                let ping = {
-                    let (connection, _) = connection.as_mut().expect("broker connection exists");
-                    sqlx::query_scalar::<_, i32>("SELECT 1").fetch_one(connection).await
-                };
-                if ping.is_err() {
-                    invalidate_battle_pet_lock_broker_like_cpp(
-                        &mut connection,
-                        &mut held_locks,
-                        &epoch,
-                    );
-                }
-            }
-        }
-    }
 }
 
 impl LoginBattlePetPersistenceLikeCpp {
@@ -932,24 +661,6 @@ impl BattlePetPersistenceLikeCpp for LoginBattlePetPersistenceLikeCpp {
     }
 }
 
-async fn load_pet_rows_like_cpp(
-    tx: &mut Transaction<'_, MySql>,
-    account_id: u32,
-    realm_id: u16,
-) -> Result<Vec<DurableBattlePetRowLikeCpp>, BattlePetPersistenceErrorLikeCpp> {
-    let rows = sqlx::query(
-        "SELECT bp.guid, bp.species, bp.breed, bp.displayId, bp.level, bp.exp, bp.health, bp.quality, bp.flags, bp.name, bp.nameTimestamp, bp.owner, dn.genitive, dn.dative, dn.accusative, dn.instrumental, dn.prepositional FROM battle_pets bp LEFT JOIN battle_pet_declinedname dn ON bp.guid = dn.guid WHERE bp.battlenetAccountId = ? AND (bp.ownerRealmId IS NULL OR bp.ownerRealmId = ?)",
-    )
-    .bind(account_id)
-    .bind(realm_id)
-    .fetch_all(&mut **tx)
-    .await
-    .map_err(|error| database_error_like_cpp(DatabaseError::from(error)))?;
-    rows.into_iter()
-        .map(|row| durable_pet_from_row_like_cpp(&row))
-        .collect()
-}
-
 async fn lock_and_validate_account_fence_like_cpp(
     tx: &mut Transaction<'_, MySql>,
     account_id: u32,
@@ -968,260 +679,3 @@ async fn lock_and_validate_account_fence_like_cpp(
         Err(BattlePetPersistenceErrorLikeCpp::StaleAuthority)
     }
 }
-
-async fn load_slot_rows_like_cpp(
-    tx: &mut Transaction<'_, MySql>,
-    account_id: u32,
-) -> Result<Vec<DurableBattlePetSlotLikeCpp>, BattlePetPersistenceErrorLikeCpp> {
-    let rows = sqlx::query(
-        "SELECT id, battlePetGuid, locked FROM battle_pet_slots WHERE battlenetAccountId = ?",
-    )
-    .bind(account_id)
-    .fetch_all(&mut **tx)
-    .await
-    .map_err(|error| database_error_like_cpp(DatabaseError::from(error)))?;
-    rows.into_iter()
-        .map(|row| {
-            let counter = row_u64_signed_or_unsigned_like_cpp(&row, 1)?;
-            Ok(DurableBattlePetSlotLikeCpp {
-                index: row_u8_signed_or_unsigned_like_cpp(&row, 0)?,
-                pet_guid_counter: (counter != 0).then_some(counter),
-                locked: row.try_get(2).map_err(|error| {
-                    BattlePetPersistenceErrorLikeCpp::Database(error.to_string())
-                })?,
-            })
-        })
-        .collect()
-}
-
-fn durable_pet_from_result_like_cpp(
-    result: &SqlResult,
-    offset: usize,
-) -> Result<DurableBattlePetRowLikeCpp, BattlePetPersistenceErrorLikeCpp> {
-    let missing = |column: usize| {
-        BattlePetPersistenceErrorLikeCpp::Database(format!(
-            "could not decode battle-pet result column {column}"
-        ))
-    };
-    Ok(DurableBattlePetRowLikeCpp {
-        guid_counter: result_u64_signed_or_unsigned_like_cpp(result, offset)
-            .ok_or_else(|| missing(offset))?,
-        species: result_u32_signed_or_unsigned_like_cpp(result, offset + 1)
-            .ok_or_else(|| missing(offset + 1))?,
-        breed: result_u16_signed_or_unsigned_like_cpp(result, offset + 2)
-            .ok_or_else(|| missing(offset + 2))?,
-        display_id: result_u32_signed_or_unsigned_like_cpp(result, offset + 3)
-            .ok_or_else(|| missing(offset + 3))?,
-        level: result_u16_signed_or_unsigned_like_cpp(result, offset + 4)
-            .ok_or_else(|| missing(offset + 4))?,
-        exp: result_u16_signed_or_unsigned_like_cpp(result, offset + 5)
-            .ok_or_else(|| missing(offset + 5))?,
-        health: result_u32_signed_or_unsigned_like_cpp(result, offset + 6)
-            .ok_or_else(|| missing(offset + 6))?,
-        quality: result_u8_signed_or_unsigned_like_cpp(result, offset + 7)
-            .ok_or_else(|| missing(offset + 7))?,
-        flags: result_u16_signed_or_unsigned_like_cpp(result, offset + 8)
-            .ok_or_else(|| missing(offset + 8))?,
-        name: result
-            .try_read(offset + 9)
-            .ok_or_else(|| missing(offset + 9))?,
-        name_timestamp: result
-            .try_read(offset + 10)
-            .ok_or_else(|| missing(offset + 10))?,
-        owner_guid_counter: if result.is_null(offset + 11) {
-            None
-        } else {
-            Some(
-                result_u64_signed_or_unsigned_like_cpp(result, offset + 11)
-                    .ok_or_else(|| missing(offset + 11))?,
-            )
-        },
-        declined_names: None,
-    })
-}
-
-async fn find_request_in_tx_like_cpp(
-    tx: &mut Transaction<'_, MySql>,
-    request_key: BattlePetAddRequestKeyLikeCpp,
-) -> Result<Option<(u32, DurableBattlePetRowLikeCpp, bool)>, BattlePetPersistenceErrorLikeCpp> {
-    let row = sqlx::query(
-        "SELECT req.battlenetAccountId, req.battlePetGuid, req.species, req.breed, req.displayId, req.level, req.exp, req.health, req.quality, req.flags, req.name, req.nameTimestamp, req.owner, pet.guid IS NOT NULL FROM battle_pet_add_requests req LEFT JOIN battle_pets pet ON pet.guid = req.battlePetGuid WHERE req.requestKey = ?",
-    )
-    .bind(request_key.as_bytes().as_slice())
-    .fetch_optional(&mut **tx)
-    .await
-    .map_err(|error| database_error_like_cpp(DatabaseError::from(error)))?;
-    let Some(row) = row else {
-        return Ok(None);
-    };
-    let account_id = row_u32_signed_or_unsigned_like_cpp(&row, 0)?;
-    let pet = DurableBattlePetRowLikeCpp {
-        guid_counter: row_u64_signed_or_unsigned_like_cpp(&row, 1)?,
-        species: row_u32_signed_or_unsigned_like_cpp(&row, 2)?,
-        breed: row_u16_signed_or_unsigned_like_cpp(&row, 3)?,
-        display_id: row_u32_signed_or_unsigned_like_cpp(&row, 4)?,
-        level: row_u16_signed_or_unsigned_like_cpp(&row, 5)?,
-        exp: row_u16_signed_or_unsigned_like_cpp(&row, 6)?,
-        health: row_u32_signed_or_unsigned_like_cpp(&row, 7)?,
-        quality: row_u8_signed_or_unsigned_like_cpp(&row, 8)?,
-        flags: row_u16_signed_or_unsigned_like_cpp(&row, 9)?,
-        name: row.try_get(10).map_err(row_decode_error_like_cpp)?,
-        name_timestamp: row.try_get(11).map_err(row_decode_error_like_cpp)?,
-        owner_guid_counter: row_opt_u64_signed_or_unsigned_like_cpp(&row, 12)?,
-        declined_names: None,
-    };
-    let still_present = row.try_get(13).map_err(row_decode_error_like_cpp)?;
-    Ok(Some((account_id, pet, still_present)))
-}
-
-fn durable_pet_from_row_like_cpp(
-    row: &sqlx::mysql::MySqlRow,
-) -> Result<DurableBattlePetRowLikeCpp, BattlePetPersistenceErrorLikeCpp> {
-    let genitive: Option<String> = row.try_get(12).map_err(row_decode_error_like_cpp)?;
-    Ok(DurableBattlePetRowLikeCpp {
-        guid_counter: row_u64_signed_or_unsigned_like_cpp(row, 0)?,
-        species: row_u32_signed_or_unsigned_like_cpp(row, 1)?,
-        breed: row_u16_signed_or_unsigned_like_cpp(row, 2)?,
-        display_id: row_u32_signed_or_unsigned_like_cpp(row, 3)?,
-        level: row_u16_signed_or_unsigned_like_cpp(row, 4)?,
-        exp: row_u16_signed_or_unsigned_like_cpp(row, 5)?,
-        health: row_u32_signed_or_unsigned_like_cpp(row, 6)?,
-        quality: row_u8_signed_or_unsigned_like_cpp(row, 7)?,
-        flags: row_u16_signed_or_unsigned_like_cpp(row, 8)?,
-        name: row.try_get(9).map_err(row_decode_error_like_cpp)?,
-        name_timestamp: row.try_get(10).map_err(row_decode_error_like_cpp)?,
-        owner_guid_counter: row_opt_u64_signed_or_unsigned_like_cpp(row, 11)?,
-        declined_names: match genitive {
-            None => None,
-            Some(genitive) => Some(BattlePetDeclinedNamesLikeCpp {
-                names: [
-                    genitive,
-                    row.try_get(13).map_err(row_decode_error_like_cpp)?,
-                    row.try_get(14).map_err(row_decode_error_like_cpp)?,
-                    row.try_get(15).map_err(row_decode_error_like_cpp)?,
-                    row.try_get(16).map_err(row_decode_error_like_cpp)?,
-                ],
-            }),
-        },
-    })
-}
-
-fn add_request_matches_like_cpp(
-    requested: &DurableBattlePetRowLikeCpp,
-    persisted: &DurableBattlePetRowLikeCpp,
-) -> bool {
-    requested.species == persisted.species
-        && requested.breed == persisted.breed
-        && requested.display_id == persisted.display_id
-        && requested.level == persisted.level
-        && requested.quality == persisted.quality
-        && requested.owner_guid_counter == persisted.owner_guid_counter
-}
-
-fn database_error_like_cpp(error: DatabaseError) -> BattlePetPersistenceErrorLikeCpp {
-    BattlePetPersistenceErrorLikeCpp::Database(error.to_string())
-}
-
-fn row_decode_error_like_cpp(error: sqlx::Error) -> BattlePetPersistenceErrorLikeCpp {
-    BattlePetPersistenceErrorLikeCpp::Database(error.to_string())
-}
-
-/// Tolerant unsigned readers for the legacy battle-pet columns (#175).
-///
-/// The legacy C++ schema declares these columns signed
-/// (`sql/base/auth_database.sql` in woltk-trinity-legacy), while the Rust
-/// structs use unsigned types and sqlx rejects the conversion outright. C++
-/// reads the same signed columns into `uint64`/`uint32`/`uint16`/`uint8`
-/// fields without caring about the declared sign. Mirror that: try the exact
-/// unsigned type first (covers rustycore-migrated unsigned schemas), then
-/// fall back to the signed column type with a range check.
-macro_rules! battle_pet_signed_or_unsigned_readers_like_cpp {
-    ($cast_fn:ident, $row_fn:ident, $result_fn:ident, $unsigned:ty, $signed:ty) => {
-        /// Checked signed→unsigned conversion shared by both tolerant readers.
-        fn $cast_fn(
-            raw: $signed,
-            column: usize,
-        ) -> Result<$unsigned, BattlePetPersistenceErrorLikeCpp> {
-            <$unsigned>::try_from(raw).map_err(|_| {
-                BattlePetPersistenceErrorLikeCpp::Database(format!(
-                    "negative value {raw} in battle-pet column {column}"
-                ))
-            })
-        }
-
-        fn $row_fn(
-            row: &sqlx::mysql::MySqlRow,
-            column: usize,
-        ) -> Result<$unsigned, BattlePetPersistenceErrorLikeCpp> {
-            if let Ok(value) = row.try_get::<$unsigned, _>(column) {
-                return Ok(value);
-            }
-            let raw: $signed = row.try_get(column).map_err(row_decode_error_like_cpp)?;
-            $cast_fn(raw, column)
-        }
-
-        fn $result_fn(result: &SqlResult, column: usize) -> Option<$unsigned> {
-            result.try_read::<$unsigned>(column).or_else(|| {
-                result
-                    .try_read::<$signed>(column)
-                    .and_then(|raw| $cast_fn(raw, column).ok())
-            })
-        }
-    };
-}
-
-battle_pet_signed_or_unsigned_readers_like_cpp!(
-    battle_pet_column_i64_as_u64_like_cpp,
-    row_u64_signed_or_unsigned_like_cpp,
-    result_u64_signed_or_unsigned_like_cpp,
-    u64,
-    i64
-);
-battle_pet_signed_or_unsigned_readers_like_cpp!(
-    battle_pet_column_i32_as_u32_like_cpp,
-    row_u32_signed_or_unsigned_like_cpp,
-    result_u32_signed_or_unsigned_like_cpp,
-    u32,
-    i32
-);
-battle_pet_signed_or_unsigned_readers_like_cpp!(
-    battle_pet_column_i16_as_u16_like_cpp,
-    row_u16_signed_or_unsigned_like_cpp,
-    result_u16_signed_or_unsigned_like_cpp,
-    u16,
-    i16
-);
-battle_pet_signed_or_unsigned_readers_like_cpp!(
-    battle_pet_column_i8_as_u8_like_cpp,
-    row_u8_signed_or_unsigned_like_cpp,
-    result_u8_signed_or_unsigned_like_cpp,
-    u8,
-    i8
-);
-
-/// Nullable variant of `row_u64_signed_or_unsigned_like_cpp` (`battle_pets.owner`).
-fn row_opt_u64_signed_or_unsigned_like_cpp(
-    row: &sqlx::mysql::MySqlRow,
-    column: usize,
-) -> Result<Option<u64>, BattlePetPersistenceErrorLikeCpp> {
-    match row.try_get::<Option<u64>, _>(column) {
-        Ok(value) => Ok(value),
-        Err(_) => {
-            let raw: Option<i64> = row.try_get(column).map_err(row_decode_error_like_cpp)?;
-            raw.map(|value| battle_pet_column_i64_as_u64_like_cpp(value, column))
-                .transpose()
-        }
-    }
-}
-
-fn is_duplicate_key_like_cpp(error: &DatabaseError) -> bool {
-    matches!(
-        error,
-        DatabaseError::Query(sqlx::Error::Database(database_error))
-            if database_error.code().as_deref() == Some("1062")
-    )
-}
-
-#[cfg(test)]
-#[path = "account_adapter/tests/mod.rs"]
-mod tests;
