+++ crates/wow-world/src/map_manager/runtime/mod.rs
@@ -1,0 +1,12 @@
+// Copyright (c) 2026 alseif0x
+// RustyCore — WoW WotLK 3.4.3 server in Rust
+// Based on TrinityCore protocol research (https://github.com/TrinityCore/TrinityCore)
+// Licensed under GPL v3 — https://www.gnu.org/licenses/gpl-3.0.html
+
+//! Legacy creature runtime frame and its per-map orchestration.
+
+use super::*;
+
+mod creature;
+mod instance;
+mod manager;
